+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## Unreleased
 ### Added
 - Added getLinearConsIndicator
+- Added SCIP_LPPARAM, setIntParam, setRealParam, getIntParam, getRealParam for lpi
+- Added isFeasPositive
 ### Fixed
 ### Changed
 ### Removed